Rename smart_stream module to smartstream
