+++ crates/fluvio-spu/src/smartstream/file_batch.rs
