+++ crates/fluvio-spu/src/smartstream/memory.rs
